Typing Test (done)/requirements.txt
pytesseract
pyautogui
pynput
Pillow